Fixed issue with user Meta
This fixes the issue when you register a new user meta, if the value is not already in the database, the GraphQL call to fetch a user would crash if that field was not set. It would get a message like "unexpected stdClass, expecting string". This is no longer the case, the object is being looked at properly and now we setup all registered fields, set their value if any or set the default for it's type.